services/trux_guardian/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
pyotp==2.9.0
cryptography==45.0.6
PyMySQL==1.1.1

